Added due date column for attention table
Each interactionstep can implement a method get_due_date() on default null is returned.
