+++ classes/table/interaction_table.php
@@ -35,17 +35,10 @@ require_once($CFG->libdir . '/tablelib.php');
 
 abstract class interaction_table extends \table_sql {
 
-    public function init() {
-        $this->define_columns(['courseid', 'courseshortname', 'coursefullname', 'status', 'tools']);
-        $this->define_headers([
-            get_string('course'),
-            get_string('shortnamecourse'),
-            get_string('fullnamecourse'),
-            get_string('status', 'tool_cleanupcourses'),
-            get_string('tools', 'tool_cleanupcourses'),
-        ]);
-        $this->setup();
-    }
+    /**
+     * Initialises the columns of the table. Necessary since attention_table has extra column date.
+     */
+    public abstract function init();
 
     /**
      * Render courseid column.
